Completed code with buttons. Next, refactor
https://www.makeschool.com/online-courses/tutorials/javascript-concentration-game/javascript-game-tutorial

Challenges
The game has some areas where it can be improved.
Over all the style is not very good. Improve the appearance of the game.
Improve the style of the squares.
Work on the interface the reset button needs some help.
The square-n styles for the colors could include images.
Improve the game play
Try adding a timer. Try and match all squares before time runs out.
Add a move count. Every click on a square counts as one move.
Add more matching elements.
Add an option that adds a larger number of squares to match.